Revert "fix: wrong type"
This reverts commit 282d7c0c007b41b7323c58b3b450ffbf98c56917.

+++ pkg/ptp/ptp.go
@@ -120,7 +120,7 @@ func (t *Torrent) UnmarshalJSON(data []byte) error {
 	}
 
 	switch id := aux.Id.(type) {
-	case int64:
+	case float64:
 		t.Id = fmt.Sprintf("%.0f", id)
 	case string:
 		t.Id = id
